update UI rendering codes
